Renamed and changed location of create activity class
