+++ app/src/main/AndroidManifest.xml
@@ -41,7 +41,7 @@
             android:name=".view.UserProfileActivity"
             android:exported="true" />
         <activity
-            android:name=".create_activity"
+            android:name=".view.CreateActivity"
             android:exported="true">
         </activity>
     </application>
