+++ app/src/main/java/com/soen490chrysalis/papilio/view/ActivitiesFragment.kt
@@ -6,7 +6,6 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.soen490chrysalis.papilio.create_activity
 import com.soen490chrysalis.papilio.databinding.FragmentActivitiesBinding
 import kotlinx.android.synthetic.main.fragment_activities.view.*
