+++ app/src/main/res/layout/activity_create_activity.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".create_activity">
+    tools:context=".view.create_activity">
 
     <FrameLayout
         android:layout_width="match_parent"
